Add api_calls to shell out to stylus' javascript api.
I needed this method to create variables for custom theming. It's essentially just a simple wrapper that shells out to commands on the stylus object. Works as a class method on Stylus, e.g:

`Stylus.api "define", "mycolor", "red"`

which uses the `define` api to create a `mycolor` global variable.

Also included a passing test.  I'd like to see this in the next ruby-stylus version in some form or another, so let me know what you think!